wip -- refactor, working send/recv file
