+++ package.json
@@ -5,14 +5,16 @@
   "description": "An Ionic project",
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "serve:worker": "vue-cli-service build --dest ./public/worker/ --target lib --formats umd --name worker --filename index --mode developement --watch ./src/worker.ts",
+    "build": "yarn build:worker && yarn build:wasm && vue-cli-service build",
+    "build:worker": "vue-cli-service build --dest ./dest/worker/ --target lib --formats umd --name worker --filename index --mode production ./src/worker.ts",
+    "build:wasm": "node scripts/build-wasm.js",
     "test:unit": "vue-cli-service test:unit",
     "test:e2e": "vue-cli-service test:e2e",
     "lint": "vue-cli-service lint",
     "deploy:s3:playground": "npm run build && aws s3 sync ./dist s3://wormhole.bryanchriswhite.com ",
     "deploy:s3:staging": "npm run build && aws s3 sync ./dist s3://staging.transfer.leastauthority.com ",
     "deploy:s3:production": "aws s3 sync ./dist s3://transfer.leastauthority.com ",
-    "build:wasm": "node scripts/build-wasm.js",
     "compose": "bash scripts/docker-compose.sh",
     "compose:logs": "bash scripts/docker-compose.sh logs -f",
     "compose:up": "bash scripts/docker-compose.sh up"
@@ -31,6 +33,7 @@
     "@capacitor/cli": "2.4.5",
     "@types/jest": "^24.0.19",
     "@types/node": "^14.14.25",
+    "@types/streamsaver": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
     "@typescript-eslint/parser": "^2.33.0",
     "@vue/cli-plugin-babel": "^4.5.11",
